Render route in AR
to be tested

+++ app/src/main/java/com/example/pbrg_android/data/model/HoldData.kt
@@ -1,6 +1,6 @@
 package com.example.pbrg_android.data.model
 
 data class HoldData(
-    val x: Int,
-    val y: Int
+    val x: Float,
+    val y: Float
 )
